requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
db-dtypes==1.2.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.7.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.48.2
idna==3.6
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.0.116
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==0.27.2
packaging==23.2
pandas==2.2.1
pandas-gbq==0.22.0
pillow==10.2.0
proto-plus==1.23.0
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.14
pydata-google-auth==1.8.2
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==1.4.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
six==1.16.0
smmap==5.0.1
SQLAlchemy==1.4.52
streamlit==1.32.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
yarl==1.9.4
protobuf==3.20.*

# # pip install protobuf==3.20.*
# langchain==0.0.116
# openai==0.27.2
# google-cloud-bigquery==3.7.0
# jinja2==3.1.2
# pandas
# python-dotenv
# db-dtypes
